Modify aevol_run option -n (now -e) and default value

+++ src/libaevol/OutputManager.h
@@ -122,7 +122,7 @@ class OutputManager {
     void write_current_generation_outputs( void ) const;
     inline void flush( void );
 
-    static int32_t get_last_gener();
+    static int64_t get_last_gener();
 
   protected :
     // =================================================================
